requirements.txt
imutils
argparse
numpy==1.22.3
matplotlib==3.6.2
opencv_python==4.7.0.68
Pillow==9.0.1
scikit_learn==1.2.2
scipy==1.9.1
tqdm==4.64.1

